\name{draw_table}
\alias{draw_table}
\title{draw_table}
\usage{draw_table(setid, RCOres_l)}
\description{draw a table with results for a specified settings set}
\arguments{
  \item{setid}{settings set ID}
  \item{RCOres_l}{Optimization result list}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
