% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_predict}
\alias{ensemble_predict}
\title{Ensemble Prediction}
\usage{
ensemble_predict(model_list, x)
}
\arguments{
\item{model_list}{List of trained networks}

\item{x}{Input features (matrix)}
}
\value{
List with mean prediction and variance (uncertainty)
}
\description{
Get predictions from all models in the ensemble.
}
