\name{f.matchCOVtoTickers}
\alias{f.matchCOVtoTickers}
\title{f.matchCOVtoTickers}
\usage{f.matchCOVtoTickers(Tickers,COV)}
\description{Make sure CovMa has the same tickers as Tickers. Make sure the sequence is also the same}
\arguments{
  \item{Tickers}{a date to be checked}
  \item{COV}{input format for a date}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
