\name{f.writeOptdetails2xlsx}
\alias{f.writeOptdetails2xlsx}
\title{f.writeOptdetails2xlsx}
\usage{f.writeOptdetails2xlsx(RCOres_l,
                              Portfolio,
                              Calcdate="",
                              ShortIndexWithOptimCash=0,
                              ZeroNetInvestment=FALSE,
                              OnlyID_v=NULL)}
\description{Find the newest Covariance Matrix in the DB that is calculated with the chosen Returns and by the chosen Calculation Method.}
\arguments{
  \item{RCOres_l}{optimization results}
  \item{Portfolio}{portfolio names}
  \item{Calcdate}{calculation date}
  \item{ShortIndexWithOptimCash}{todo}
  \item{ZeroNetInvestment}{is leverage or cash allowed}
  \item{OnlyID_v}{which Covariance matrix IDs to use}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
