% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{generate_architectures}
\alias{generate_architectures}
\title{Generate Random Architecture}
\usage{
generate_architectures(n)
}
\arguments{
\item{n}{Number of architectures to generate}
}
\value{
A data frame with architecture specifications
}
\description{
Generates a random neural network architecture specification.
}
\keyword{internal}
