\name{get_ap_bb_ticker_by_name}
\alias{get_ap_bb_ticker_by_name}
\title{get_ap_bb_ticker_by_name}
\usage{get_ap_bb_ticker_by_name(portfolioName, connection)}
\description{The function returns a ticker name from DB}
\arguments{
  \item{portfolioName}{A portfolio name}
  \item{connection}{Database connection to use}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
