\name{get_sh_cov_run_ids}
\alias{get_sh_cov_run_ids}
\title{get_sh_cov_run_ids}
\usage{get_sh_cov_run_ids(portfolio_name)}
\description{short tickers for printing in plots}
\arguments{
  \item{portfolio_name}{Portfolio Name for which covariance matrices are searched}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
