\name{OptimizationResultsCharacteristics}
\alias{OptimizationResultsCharacteristics}
\title{OptimizationResultsCharacteristics}
\usage{OptimizationResultsCharacteristics(rw,cov,lb,ub,set,rb_target)}
\description{The function checks if the date is in right format}
\arguments{
  \item{rw}{todo}
  \item{cov}{covariance matrix}
  \item{lb}{lower bound}
  \item{ub}{upper bound}
  \item{set}{input options}
  \item{rb_target}{risk budget targets}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
