\name{plotSingleSetIDfromRCOres}
\alias{plotSingleSetIDfromRCOres}
\title{plotSingleSetIDfromRCOres}
\usage{plotSingleSetIDfromRCOres(setid, RCOres_l)}
\description{plot optimization results for a specified settings set}
\arguments{
  \item{setid}{settings set ID}
  \item{RCOres_l}{Optimization result list}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
