% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps.R
\name{print.aps_reliability}
\alias{print.aps_reliability}
\title{Print reliability ratio diagnostic}
\usage{
\method{print}{aps_reliability}(x, ...)
}
\arguments{
\item{x}{An aps_reliability object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print reliability ratio diagnostic
}
