% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps.R
\name{print.aps_result}
\alias{print.aps_result}
\title{Print APS Result}
\usage{
\method{print}{aps_result}(x, ...)
}
\arguments{
\item{x}{An aps_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print APS Result
}
