% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps.R
\name{summary.aps_result}
\alias{summary.aps_result}
\title{Summary of APS Result}
\usage{
\method{summary}{aps_result}(object, ...)
}
\arguments{
\item{object}{An aps_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary of APS Result
}
