% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{train_ensemble}
\alias{train_ensemble}
\title{Train Deep Ensemble}
\usage{
train_ensemble(x_train, y_train, architecture_df, verbose = TRUE, ...)
}
\arguments{
\item{x_train}{Training features (matrix)}

\item{y_train}{Training targets (vector)}

\item{architecture_df}{Data frame specifying architectures}

\item{verbose}{Print progress}

\item{...}{Additional arguments passed to train_network}
}
\value{
List of trained networks
}
\description{
Trains an ensemble of neural networks with varying architectures.
}
\keyword{internal}
