% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{train_single_model}
\alias{train_single_model}
\title{Train a Single Model (Helper for Parallelization)}
\usage{
train_single_model(arch_row, x_train, y_train, n_params, ...)
}
\arguments{
\item{arch_row}{Single row from architecture data frame}

\item{x_train}{Training features}

\item{y_train}{Training targets}

\item{n_params}{Number of input parameters}

\item{...}{Additional arguments passed to train_network}
}
\value{
Trained network
}
\description{
Creates and trains a single neural network. Used internally for parallel training.
}
\keyword{internal}
