\name{update_default_settings}
\alias{update_default_settings}
\title{update_default_settings}
\usage{update_default_settings(pname, setID, covID)}
\description{plots optimization results}
\arguments{
  \item{pname}{portfolio name}
  \item{setID}{new default settings ID}
  \item{covID}{covariance matrix ID}
}
%\examples{isDate(mydate, format = "\%Y-\%m-\%d")}
