% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aps.R
\name{evaluate_objective}
\alias{evaluate_objective}
\title{Evaluate Objective Function}
\usage{
evaluate_objective(obj_function, x, num_cores = 1)
}
\arguments{
\item{obj_function}{The objective function}

\item{x}{Matrix of parameter values}

\item{num_cores}{Number of cores}
}
\value{
Vector of performance statistics
}
\description{
Helper function to evaluate the objective function, optionally in parallel.
}
\keyword{internal}
